+++ domains/certificates/Query.API/API.UnitTests/GranularCertificateIssuer/EnergyMeasuredEventHandlerTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using API.ContractService;
+using API.Data;
 using API.GranularCertificateIssuer;
 using CertificateValueObjects;
 using FluentAssertions;
