+++ domains/certificates/Query.API/API/ContractService/Repositories/CertificateIssuingContractRepository.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using API.Data;
 
 namespace API.ContractService.Repositories;
