+++ domains/certificates/Query.API/API/Data/ApplicationDbContext.cs
@@ -1,0 +1,27 @@
+using API.ContractService;
+using API.DataSyncSyncer.Persistence;
+using Microsoft.EntityFrameworkCore;
+
+namespace API.Data;
+
+//TODO: What about location of migration scripts?
+public class ApplicationDbContext : DbContext
+{
+    public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
+    {
+    }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Entity<CertificateIssuingContract>().HasIndex(c => new { c.GSRN, c.ContractNumber }).IsUnique();
+
+        modelBuilder.Entity<ProductionCertificate>().OwnsOne(c => c.Technology);
+        modelBuilder.Entity<ProductionCertificate>().HasIndex(c => new { c.Gsrn, c.DateFrom, c.DateTo }).IsUnique();
+
+        modelBuilder.Entity<SynchronizationPosition>().HasKey(s => s.GSRN);
+    }
+
+    public DbSet<CertificateIssuingContract> Contracts { get; set; }
+    public DbSet<ProductionCertificate> ProductionCertificates { get; set; }
+    public DbSet<SynchronizationPosition> SynchronizationPositions { get; set; }
+}
