+++ domains/certificates/Query.API/API/DataSyncSyncer/DataSyncSyncerWorker.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using API.ContractService;
+using API.Data;
 using API.DataSyncSyncer.Client.Dto;
 using MassTransit;
 using MeasurementEvents;
