+++ domains/certificates/Query.API/API/DataSyncSyncer/Persistence/SyncState.cs
@@ -2,6 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
+using API.Data;
 
 namespace API.DataSyncSyncer.Persistence;
