+++ domains/certificates/Query.API/API/GranularCertificateIssuer/EnergyMeasuredEventHandler.cs
@@ -1,6 +1,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 using API.ContractService;
+using API.Data;
 using CertificateValueObjects;
 using Contracts.Certificates;
 using MassTransit;
