+++ domains/certificates/Query.API/API/GranularCertificateIssuer/Startup.cs
@@ -1,3 +1,4 @@
+using API.Data;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
