+++ domains/certificates/Query.API/API/Program.cs
@@ -20,6 +20,7 @@ using Serilog.Formatting.Json;
 using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
+using API.Data;
 
 var builder = WebApplication.CreateBuilder(args);
